refactor: features 삭제
