+++ pages/login.tsx
@@ -8,7 +8,7 @@ import { type PostLoginProps } from '@/apis/auth/postLogin'
 import { type ServerErrorRes } from '@/apis/client'
 import Button from '@/components/Button'
 import Input from '@/components/Input'
-import { useLogin } from '@/features/auth/mutations/useLogin'
+import { useLogin } from '@/hooks/auth/mutations/useLogin'
 import { classNames } from '@/utils/classNames'
 
 const NotoSans = Noto_Sans({
