+++ pages/register.tsx
@@ -7,10 +7,10 @@ import DatePickerField from '@/components/DatepickerField'
 import Input from '@/components/Input'
 import InputLabel from '@/components/InputLabel'
 import Navbar from '@/components/Navbar'
-import { useCheckLoginId } from '@/features/auth/mutations/useCheckLoginId'
-import { useEmailCode } from '@/features/auth/mutations/useEmailCode'
-import { useSignUp } from '@/features/auth/mutations/useSignUp'
-import { useVerifyEmail } from '@/features/auth/mutations/useVerifyEmail'
+import { useCheckLoginId } from '@/hooks/auth/mutations/useCheckLoginId'
+import { useEmailCode } from '@/hooks/auth/mutations/useEmailCode'
+import { useSignUp } from '@/hooks/auth/mutations/useSignUp'
+import { useVerifyEmail } from '@/hooks/auth/mutations/useVerifyEmail'
 import { classNames } from '@/utils/classNames'
 
 const NotoSans = Noto_Sans({
